Funcs moved to BasicFuns
