+++ modules/index.d.ts
@@ -841,4 +841,5 @@ export const Utils: Utils;
 export const Query: Query;
 export const Builder: Builder;
 export const BasicConfig: BasicConfig;
+export const BasicFuncs: Funcs;
 export const Widgets: ReadyWidgets;
