+++ modules/index.js
@@ -6,8 +6,9 @@ import * as Import from "./import";
 import * as Widgets from "./components/widgets/index";
 import * as Operators from "./components/operators";
 import * as BasicUtils from "./utils";
+import * as BasicFuncs from "./config/funcs";
 const Utils = {...BasicUtils, ...Export, ...Import};
-export {Widgets, Operators, Utils, Export, Import};
+export {Widgets, Operators, Utils, Export, Import, BasicFuncs};
 
 export {default as BasicConfig} from "./config/basic";
